+++ src/components/Builder/Builder.tsx
@@ -1,7 +1,9 @@
 import { useCallback } from 'react';
 import Button from 'semantic-ui-react/dist/commonjs/elements/Button';
+import Tab from 'semantic-ui-react/dist/commonjs/modules/Tab';
 import { PokemonSelector } from 'common';
 import { Member } from 'components';
+import Coverage from 'components/Builder/elements/Coverage/Coverage';
 import useStore from 'store';
 import { ReactComponent as PokeballSVG } from 'assets/svg/pokeball.svg';
 import styles from './Builder.module.scss';
@@ -18,38 +20,56 @@ function Builder(): JSX.Element {
   const handlePokemon = (id: number) => {
     addTeamMember(id);
   };
-  return (
-    <div className={styles.builder}>
-      {team?.length > 0 ? (
-        <div className={styles.team}>
-          {team?.map((poke, i) => {
-            return <Member index={i} key={`team-${poke.id}`} pokemonDetail={poke} />;
-          })}
-        </div>
-      ) : (
-        <div className={styles.placeholder}>
-          <PokeballSVG />
-          {selectedGame ? (
-            <span>Press + to begin building a team</span>
-          ) : (
-            <span>Please select a game</span>
-          )}
-        </div>
-      )}
-      {selectedGame && (
-        <PokemonSelector handlePokemon={handlePokemon}>
-          <Button
-            circular
-            className={styles.button}
-            data-testid="builder-add"
-            disabled={team?.length === 6}
-            icon="plus"
-            type="button"
-          />
-        </PokemonSelector>
-      )}
-    </div>
-  );
+
+  const panes = [
+    {
+      menuItem: 'Builder',
+      render: () => (
+        <Tab.Pane attached={false}>
+          <div className={styles.builder}>
+            {team?.length > 0 ? (
+              <div className={styles.team}>
+                {team?.map((poke, i) => {
+                  return <Member index={i} key={`team-${poke.id}`} pokemonDetail={poke} />;
+                })}
+              </div>
+            ) : (
+              <div className={styles.placeholder}>
+                <PokeballSVG />
+                {selectedGame ? (
+                  <span>Press + to begin building a team</span>
+                ) : (
+                  <span>Please select a game</span>
+                )}
+              </div>
+            )}
+            {selectedGame && (
+              <PokemonSelector handlePokemon={handlePokemon}>
+                <Button
+                  circular
+                  className={styles.button}
+                  data-testid="builder-add"
+                  disabled={team?.length === 6}
+                  icon="plus"
+                  type="button"
+                />
+              </PokemonSelector>
+            )}
+          </div>
+        </Tab.Pane>
+      ),
+    },
+    {
+      menuItem: 'Coverage',
+      render: () => (
+        <Tab.Pane attached={false}>
+          <Coverage />
+        </Tab.Pane>
+      ),
+    },
+  ];
+
+  return <Tab className={styles.tabs} menu={{ secondary: true }} panes={panes} />;
 }
 
 export default Builder;