+++ src/components/Builder/elements/Coverage/Coverage.tsx
@@ -1,0 +1,19 @@
+// import { selectBuilderStrong, selectBuilderWeak } from 'selectors';
+// import useStore from 'store';
+// import styles from './Coverage.module.scss';
+
+function Coverage(): JSX.Element {
+  // const strongArr = useStore(selectBuilderStrong);
+  // const weakArr = useStore(selectBuilderWeak);
+  // console.log(strongArr);
+  // console.log(weakArr);
+  return (
+    <div style={{ color: 'var(--contrast)' }}>
+      {/* <div>Strong against</div>
+      <div>Weak Against</div> */}
+      Coming soon
+    </div>
+  );
+}
+
+export default Coverage;